PAL-999 checkstyle
