+++ data-service/src/contract-tests/java/uk/gov/gchq/palisade/contract/data/kafka/KafkaInitializer.java
@@ -14,17 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.palisade.contract.resource.kafka;
+package uk.gov.gchq.palisade.contract.data.kafka;
 
 import akka.actor.ActorSystem;
 import akka.stream.Materializer;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.NewTopic;
-import org.apache.kafka.common.serialization.Deserializer;
-import org.apache.kafka.common.serialization.Serializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -35,13 +30,11 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.core.env.Environment;
 import org.springframework.core.io.ResourceLoader;
-import org.springframework.core.serializer.support.SerializationFailedException;
 import org.springframework.test.context.support.TestPropertySourceUtils;
 import org.testcontainers.containers.KafkaContainer;
 
-import uk.gov.gchq.palisade.service.resource.stream.PropertiesConfigurer;
+import uk.gov.gchq.palisade.service.data.stream.PropertiesConfigurer;
 
-import java.io.IOException;
 import java.util.AbstractMap;
 import java.util.List;
 import java.util.Map;
@@ -53,14 +46,12 @@ import static org.apache.kafka.clients.admin.AdminClientConfig.BOOTSTRAP_SERVERS
 
 public class KafkaInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
     private static final Logger LOGGER = LoggerFactory.getLogger(KafkaInitializer.class);
-    private static final ObjectMapper MAPPER = new ObjectMapper();
 
-
-    static final KafkaContainer KAFKA_CONTAINER = new KafkaContainer("5.5.1")
+    static final KafkaContainer KAFKA = new KafkaContainer("5.5.1")
             .withReuse(false);
 
-    static void createTopics(final List<NewTopic> newTopics, final KafkaContainer kafka) throws ExecutionException, InterruptedException {
-        try (AdminClient admin = AdminClient.create(Map.of(BOOTSTRAP_SERVERS_CONFIG, String.format("%s:%d", "localhost", kafka.getFirstMappedPort())))) {
+    static void createTopics(final List<NewTopic> newTopics) throws ExecutionException, InterruptedException {
+        try (AdminClient admin = AdminClient.create(Map.of(BOOTSTRAP_SERVERS_CONFIG, String.format("%s:%d", "localhost", KafkaInitializer.KAFKA.getFirstMappedPort())))) {
             admin.createTopics(newTopics);
             LOGGER.info("created topics: " + admin.listTopics().names().get());
         }
@@ -68,14 +59,14 @@ public class KafkaInitializer implements ApplicationContextInitializer<Configura
 
     @Override
     public void initialize(final ConfigurableApplicationContext configurableApplicationContext) {
-        configurableApplicationContext.getEnvironment().setActiveProfiles("akka-test", "db-test", "test-resource");
-        KAFKA_CONTAINER.addEnv("KAFKA_AUTO_CREATE_TOPICS_ENABLE", "false");
-        KAFKA_CONTAINER.addEnv("KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR", "1");
-        KAFKA_CONTAINER.start();
+        configurableApplicationContext.getEnvironment().setActiveProfiles("akka-test");
+        KAFKA.addEnv("KAFKA_AUTO_CREATE_TOPICS_ENABLE", "false");
+        KAFKA.addEnv("KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR", "1");
+        KAFKA.start();
 
         // test kafka config
         String kafkaConfig = "akka.discovery.config.services.kafka.from-config=false";
-        String kafkaPort = "akka.discovery.config.services.kafka.endpoints[0].port" + KAFKA_CONTAINER.getFirstMappedPort();
+        String kafkaPort = "akka.discovery.config.services.kafka.endpoints[0].port" + KAFKA.getFirstMappedPort();
         TestPropertySourceUtils.addInlinedPropertiesToEnvironment(configurableApplicationContext, kafkaConfig, kafkaPort);
     }
 
@@ -83,8 +74,7 @@ public class KafkaInitializer implements ApplicationContextInitializer<Configura
     public static class Config {
 
         private final List<NewTopic> topics = List.of(
-                new NewTopic("user", 1, (short) 1),
-                new NewTopic("resource", 1, (short) 1),
+                new NewTopic("success", 1, (short) 1),
                 new NewTopic("error", 1, (short) 1));
 
         @Bean
@@ -95,8 +85,8 @@ public class KafkaInitializer implements ApplicationContextInitializer<Configura
 
         @Bean
         KafkaContainer kafkaContainer() throws ExecutionException, InterruptedException {
-            createTopics(this.topics, KAFKA_CONTAINER);
-            return KAFKA_CONTAINER;
+            createTopics(this.topics);
+            return KAFKA;
         }
 
         @Bean
@@ -107,6 +97,7 @@ public class KafkaInitializer implements ApplicationContextInitializer<Configura
                     props.getAllActiveProperties().entrySet().stream()
                             .filter(kafkaPort -> !kafkaPort.getKey().equals("akka.discovery.config.services.kafka.endpoints[0].port")),
                     Stream.of(new AbstractMap.SimpleEntry<>("akka.discovery.config.services.kafka.endpoints[0].port", Integer.toString(kafka.getFirstMappedPort()))))
+                    .peek(entry -> LOGGER.info("Config key {} = {}", entry.getKey(), entry.getValue()))
                     .collect(toMap(Map.Entry::getKey, Map.Entry::getValue))));
         }
 
@@ -114,30 +105,6 @@ public class KafkaInitializer implements ApplicationContextInitializer<Configura
         @Primary
         Materializer materializer(final ActorSystem system) {
             return Materializer.createMaterializer(system);
-        }
-    }
-
-    // Serialiser for upstream test input
-    static class RequestSerialiser implements Serializer<JsonNode> {
-        @Override
-        public byte[] serialize(final String s, final JsonNode resourceRequest) {
-            try {
-                return MAPPER.writeValueAsBytes(resourceRequest);
-            } catch (JsonProcessingException e) {
-                throw new SerializationFailedException("Failed to serialise " + resourceRequest.toString(), e);
-            }
-        }
-    }
-
-    // Deserialiser for downstream test output
-    static class ResponseDeserialiser implements Deserializer<JsonNode> {
-        @Override
-        public JsonNode deserialize(final String s, final byte[] resourceResponse) {
-            try {
-                return MAPPER.readTree(resourceResponse);
-            } catch (IOException e) {
-                throw new SerializationFailedException("Failed to deserialise " + new String(resourceResponse), e);
-            }
         }
     }
 }
