+++ filtered-resource-service/src/contract-tests/java/uk/gov/gchq/palisade/contract/filteredresource/redis/RedisInitializer.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2018-2021 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.gchq.palisade.contract.filteredresource.redis;
+
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.lang.NonNull;
+import org.springframework.test.context.support.TestPropertySourceUtils;
+import org.testcontainers.containers.GenericContainer;
+
+public class RedisInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+    public static final int REDIS_PORT = 6379;
+
+    static final GenericContainer<?> REDIS = new GenericContainer<>("redis:6-alpine")
+            .withExposedPorts(REDIS_PORT)
+            .withReuse(true);
+
+    @Override
+    public void initialize(@NonNull final ConfigurableApplicationContext context) {
+        // Start container
+        REDIS.start();
+
+        // Override Redis configuration
+        String redisContainerIP = "spring.redis.host=" + REDIS.getContainerIpAddress();
+        // Configure the testcontainer random port
+        String redisContainerPort = "spring.redis.port=" + REDIS.getMappedPort(REDIS_PORT);
+        // Override the configuration at runtime
+        TestPropertySourceUtils.addInlinedPropertiesToEnvironment(context, redisContainerIP, redisContainerPort);
+    }
+}
