+++ user-service/src/component-tests/java/uk/gov/gchq/palisade/component/user/service/RedisInitializer.java
@@ -14,29 +14,38 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.palisade.contract.filteredresource.kafka;
+package uk.gov.gchq.palisade.component.user.service;
 
+import org.jetbrains.annotations.NotNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.lang.NonNull;
 import org.springframework.test.context.support.TestPropertySourceUtils;
 import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.Network;
 
 public class RedisInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(RedisInitializer.class);
+
     private static final int REDIS_PORT = 6379;
+
     static final GenericContainer<?> REDIS = new GenericContainer<>("redis:6-alpine")
             .withExposedPorts(REDIS_PORT)
+            .withNetwork(Network.SHARED)
             .withReuse(true);
 
     @Override
-    public void initialize(@NonNull final ConfigurableApplicationContext context) {
+    public void initialize(@NotNull final ConfigurableApplicationContext context) {
+        context.getEnvironment().setActiveProfiles("redis");
         // Start container
         REDIS.start();
 
         // Override Redis configuration
         String redisContainerIP = "spring.redis.host=" + REDIS.getContainerIpAddress();
-        // Configure the testcontainer random port
+        // Configure the test container random port
         String redisContainerPort = "spring.redis.port=" + REDIS.getMappedPort(REDIS_PORT);
+        LOGGER.info("Starting Redis with {}", redisContainerPort);
         // Override the configuration at runtime
         TestPropertySourceUtils.addInlinedPropertiesToEnvironment(context, redisContainerIP, redisContainerPort);
     }
