+++ user-service/src/contract-tests/java/uk/gov/gchq/palisade/contract/user/kafka/KafkaInitializer.java
@@ -55,7 +55,7 @@ import java.util.stream.Stream;
 class KafkaInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
 
     static final KafkaContainer KAFKA_CONTAINER = new KafkaContainer("5.5.1")
-            .withReuse(true);
+            .withReuse(false);
     private static final Logger LOGGER = LoggerFactory.getLogger(KafkaContractTest.class);
     private static final ObjectMapper MAPPER = new ObjectMapper();
